Reorganize fromLLm folder just by features
