+++ src/app/features/editor/fromLLM/llm2keml/llm-conversation-creation.service.ts
@@ -1,4 +1,4 @@
-import {LLMMessage} from "@app/features/editor/fromLLM/models/llmmessage";
+import {LLMMessage} from "@app/features/editor/fromLLM/llm2keml/llmmessage";
 import {Conversation} from "@app/shared/keml/models/core/conversation";
 import {ConversationPartner} from "@app/shared/keml/models/core/conversation-partner";
 import {Author} from "@app/shared/keml/models/core/author";
